==> procedure/code/01_population_data_tidycensus.Rmd <==
---
title: "Retrieving US Census Bureau Data (Population Dataset) "
output: html_document
date: "2024-01-16"
---

```{r environment-setup, include = FALSE}
# record all the packages you are using here
# this includes any calls to library(), require(),
# and double colons such as here::i_am()
packages <- c("tidyverse", "here", "tidycensus", "sf", "tigris")

# force all conflicts to become errors
# if you load dplyr and use filter(), R has to guess whether you mean dplyr::filter() or stats::filter()
# the conflicted package forces you to be explicit about this
# disable at your own peril
# https://conflicted.r-lib.org/
require(conflicted)

# load and install required packages
# https://groundhogr.com/
if (!require(groundhog)) {
  install.packages("groundhog")
  require(groundhog)
}

# this date will be used to determine the versions of R and your packages
# it is best practice to keep R and its packages up to date
groundhog.day <- "2025-04-01"

# this replaces any library() or require() calls
groundhog.library(packages, groundhog.day)
# you may need to install a correct version of R
# you may need to respond OK in the console to permit groundhog to install packages
# you may need to restart R and rerun this code to load installed packages
# In RStudio, restart r with Session -> Restart Session
```


```{r}
# record the R processing environment
# alternatively, use devtools::session_info() for better results
writeLines(
  capture.output(sessionInfo()),
  here("procedure", "environment", paste0("r-environment-", Sys.Date(), ".txt"))
)

# save package citations
knitr::write_bib(c(packages, "base"), file = here("software.bib"))

# set up default knitr parameters
# https://yihui.org/knitr/options/
knitr::opts_chunk$set(
  echo = FALSE, # Show outputs, but not code. Change to TRUE to show code as well
  fig.retina = 4,
  fig.width = 8,
  fig.path = paste0(here("results", "figures"), "/")
)

options(tigris_use_cache = TRUE)
conflicts_prefer(dplyr::filter)
```


```{r}
counties <- counties(state = c("VT", "NH", "MA", "NY"), year = 2020)
```


```{r}
# select counties within 40 mile buffer of Vermont
countiesVT <- counties |> 
  dplyr::filter(STATEFP == "50") |> 
  select(geometry) |> 
  st_buffer(64373)

bufferROI <- counties |> 
  dplyr::filter(STATEFP != "50") |> 
  st_join(countiesVT, join=st_intersects, left=FALSE) |> 
  select(STATEFP, COUNTYFP, GEOID) |> 
  st_drop_geometry() |> 
  distinct()
```


```{r}
vars2020 <- load_variables(2020, "dhc")
myvars <- c("P1_001N", 
            "P11_001N", "P11_005N", 
            "P12_020N", "P12_021N", "P12_022N", "P12_023N", "P12_024N", "P12_025N",
            "P12_044N", "P12_045N", "P12_046N", "P12_047N", "P12_048N", "P12_049N")

dhc_dictionary <- vars2020 |> dplyr::filter(name %in% myvars) |> 
  arrange(name)
write_csv(dhc_dictionary, here("data", "metadata", "census2020dhc_variables.csv"))

```

# Vermont Blocks

```{r}
vtBlocks <- get_decennial(
  geography = "block",
  state = "50",
  geometry = TRUE,
  year = 2020,
  sumfile = "dhc",
  variables = myvars,
  output = "wide"
)

st_write(vtBlocks, blocks50_file)
```

# Buffer Blocks

```{r}
states <- bufferROI |> select(STATEFP) |> distinct()

for (s in states$STATEFP) {
  roiCounties <- bufferROI |> 
    dplyr::filter(STATEFP == s) |> 
    select(COUNTYFP)

  countyResults <- c()  
  for (c in roiCounties) {
    curBlocks <- get_decennial(geography = "block",
                              state = s,
                              county = c,
                              geometry = TRUE,
                              year = 2020,
                              sumfile = "dhc",
                              variables = myvars,
                              output = "wide")
    countyResults <- countyResults |> append(c(curBlocks))
  }
  countyResults <- bind_rows(countyResults)
  st_write(countyResults, here("data", "raw", "public", "population", paste0("blocks", s, ".gpkg")))
}

```

## Query County Subdivisions

```{r}
states <- c("25", "33", "36")

subdivisions <- get_decennial(geography = "county subdivision",
                              state = "50",
                              geometry = TRUE,
                              year = 2020,
                              sumfile = "dhc",
                              variables = myvars,
                              output = "wide")

for (s in states) {
  roiCounties <- bufferROI |> 
    dplyr::filter(STATEFP == s) |> 
    select(COUNTYFP)
  
  for (c in roiCounties) {
    subdivision_set <- get_decennial(geography = "county subdivision",
                                      state = s,
                                      county = c,
                                      geometry = TRUE,
                                      year = 2020,
                                      sumfile = "dhc",
                                      variables = myvars,
                                      output = "wide")
    subdivisions <- bind_rows(subdivisions, subdivision_set)
  }

}

st_write(subdivisions, 
         here("data", "raw", "public", "population", "county_subdivisions.gpkg"),
         append = FALSE)
```

## Process and Save Population Data

Function to process census data

```{r}
process_census <- function(df) {
  df_derived <- df |>
    mutate(
      total_pop = P1_001N,
      minority_pop = P11_001N - P11_005N
    )
  
  df_derived$elderly_pop <- df_derived |>
    st_drop_geometry() |>
    select(starts_with("P12")) |>
    rowSums()
  
  return(df_derived)
}
```

Process county subdivisions and save

```{r}
subdivisions <- st_read(here("data", "raw", "public", "population", "county_subdivisions.gpkg"))
subdivisions_derived <- process_census(subdivisions)
st_write(subdivisions_derived, 
         here("data", "derived", "public", "population", "county_subdivisions.gpkg"),
         append = FALSE)
```

Process blocks and save

```{r}
states <- c("50", "25", "33", "36")

blockCentroids <- list()

for (s in states) {
  # load blocks for state s
  blockFiler <- here("data", "raw", "public", "population", 
                    paste0("blocks", s, ".gpkg"))
  blockFiled <- here("data", "derived", "public", "population", 
                    paste0("blocks", s, ".gpkg"))
  blocks <- st_read(blockFiler)
  
  # filter for populated blocks and calculate derived variables
  blocksDerived <- blocks |> 
    filter(P1_001N > 0) |> 
    process_census()
  
  # find point on surface and join the intersecting county subdivision ID
  blockPts <- blocksDerived |> st_point_on_surface() |> 
    st_join(select(subdivisions, sdGEOID=GEOID))
  
  # create block to subdivision crosswalk file by GEOID
  blocksdcross <- blockPts |> st_drop_geometry() |> select(GEOID, sdGEOID)
  
  blocksDerived <- blocksDerived |> left_join(blocksdcross, by="GEOID")
  
  blocksDerived |> st_write(blockFiled, append=FALSE)
  
  blockCent <- blocksDerived |> st_centroid()
  
  blockCentroids <- append(blockCentroids, list(blockCent))
}

blocks50_file <- here("data", "raw", "public", "population", "blocks50.gpkg")

blockCentroidsBind <- bind_rows(blockCentroids)

st_write(blockCentroidsBind, 
         here("data", "derived", "public", "population", "block_centroids.gpkg"),
         append = FALSE)
```
